+++ client/build.js
@@ -1,5 +1,6 @@
 const { build } = require('esbuild');
 const copyStaticFiles = require('esbuild-copy-static-files');
+const svgr =require('esbuild-plugin-svgr');
 
 
 async function runBuild() {
@@ -16,7 +17,8 @@ async function runBuild() {
                 dest: './dist',
                 dereference: true,
                 recursive: true,
-            })
+            }),
+            svgr()
         ]
     })
 }