+++ client/src/components/notification.jsx
@@ -88,7 +88,7 @@ export default () => {
                         onClick={() => { setIsOpen(false); }}
                         style={{
                             position: 'absolute',
-                            bottom: '100px'
+                            bottom: '40px'
                         }}
                     >
                         OK