+++ client/src/platformModules/clio.js
@@ -1,5 +1,9 @@
-function openContactPage(hostname, incomingCallContactInfo){
+function openContactPage(hostname, incomingCallContactInfo) {
     window.open(`https://${hostname}/nc/#/contacts/${incomingCallContactInfo.id}`);
+}
+async function onUnauthorize(){
+
 }
 
 exports.openContactPage = openContactPage;
+exports.onUnauthorize = onUnauthorize;