+++ docs/release-notes.md
@@ -6,3 +6,19 @@
 - Feature: Now support multiple matched contacts in call logging. Users can select target contact if it's a multiple match case
 - New setting - "Open contact web page after creating it": Auto open placeholder contact page so that users can do further editing more easily
 - UI: Polished the looking of some UIs
+
+## 0.8.1
+
+- Fix: Sign up button now opens RingCentral plan page
+
+## 0.8.2
+
+- Fix: Message conversation log page open issue
+- Feature: Clicking log button for a logged call will open corresponding contact/log page
+
+## 0.8.3
+
+- Change: Updated logo icon
+- Fix: SMS log issue in multiple contacts matched case
+- Fix: In-call note button will disappear if user opens message conversation so to not cover SMS send button
+- Bullhorn: Clear user session info when log out